Single practice paper service completed with all crud operations

+++ src/main/java/in/co/examsadda/crud/repository/QuestionRepository.java
@@ -1,5 +1,6 @@
 package in.co.examsadda.crud.repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
@@ -8,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import in.co.examsadda.entity.Question;
 
 @Repository
-public interface QuestionRepository extends CrudRepository<Question, Integer> {
+public interface QuestionRepository extends CrudRepository<Question, BigInteger> {
 	
 	public Question findByQuestionId(Integer questionId);
